refactor(tests): extract outline.rs (10 tests)
Move get_outline_uses_named_realm + 9 outline_* / outline_tree_* tests
from mod.rs into engine/tests/outline.rs.

Tests moved (now at engine::tests::outline::*):
  get_outline_uses_named_realm
  outline_flat_format_backward_compat
  outline_tree_format_nested_hierarchy
  outline_tree_format_skipped_levels
  outline_tree_format_no_headings
  outline_tree_root_node_no_heading
  outline_include_text_false_omits_field
  outline_include_text_true_inlines_content
  outline_structured_doc_tree_fallback_to_flat
  outline_unicode_headings

Tests: 135 passed / 0 failed / 1 ignored (baseline unchanged).

marky-n1h step 6.
